requirements.txt
# Database dependencies
psycopg2-binary
apache-airflow
apache-airflow-providers-openlineage >= 1.8.0
SQLAlchemy<2.0

# Data analysis dependencies
pandas
pytrends

# Other
requests
pydantic<=2.0
python-dotenv
pyyaml
pytest
pytest-mock